requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
keras>=2.12.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0

# Kaggle API
kaggle>=1.5.16

# Environment variables
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics and evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Progress bars
tqdm>=4.66.0

# XML parsing (for Pascal VOC annotations)
lxml>=4.9.0
